potential fix for transparent gifs
gifs are compressed by using previous frames' data to fill in current frame data, thus meaning you only need one occurance of that frame data in the file, greatly reducing file size for some images. however this also means that there is now a void there when using this to decode it, so i have to do this. y did i make a paragraph about it omg